API Complete!
The api is ready for action with a stable local host and all methods working properly(there is even some lazy documentation)! Ready to move onto the front end aspects of the project! (I will need to go back and enable cors for the front end project once I create it.